requirements.txt
# Core dependencies
discord.py>=2.3.2
python-dotenv>=1.0.0
redis>=4.6.0
aiohttp>=3.9.3
SQLAlchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
uvloop>=0.17.0; sys_platform != "win32"  # Optional performance boost for Unix systems

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
mypy>=1.5.1
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
types-setuptools>=68.0.0
types-redis>=4.6.0.5
types-aiofiles>=23.1.0.5

# Type stubs
types-python-dateutil>=2.8.19.14
types-requests>=2.31.0.2
types-urllib3>=1.26.25.14

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0

# Security
bandit>=1.7.5
safety>=2.3.5

# Formatting and linting
pylint>=2.17.5
autopep8>=2.0.4
yapf>=0.40.1
pycodestyle>=2.11.0
pydocstyle>=6.3.0

# Debug tools
ipython>=8.14.0
ipdb>=0.13.13
rich>=13.5.2
